+++ backend/channel-service/src/test/java/com/vdt2026/omnicare/channel/events/application/InboundEventDispatchServiceTests.java
@@ -50,6 +50,7 @@ class InboundEventDispatchServiceTests {
                 externalMessageId,
                 "fb-user",
                 "Facebook User",
+                null,
                 "hello",
                 occurredAt
             )